Refactor Assets
